[MeshTensor] Make tensor data-conversion utilities private to tt_metal
Reduce the public API surface of experimental/tensor/impl/tensor_impl.hpp. The data-conversion templates (encode_tensor_data, decode_tensor_data, to_tile_major_layout, to_row_major_layout) are internal-only and move to a new private header tt_metal/impl/tensor/tensor_impl_private.hpp. The public header retains only the buffer allocators (allocate_device_buffer, allocate_host_buffer) and the shared DataType dispatch toolkit (dispatch + bfloat4_b/bfloat8_b tag structs), which ttnn still uses.

External callers are migrated to existing public Runtime Tensor APIs: pytensor.cpp to_torch/to_numpy use HostTensor::to_vector and tt::tt_metal::to_layout + host_buffer::get_as; tt-train nb_util.cpp uses Tensor::to_vector; bfloat_dtype_traits.hpp drops the dead bfloat8_b/bfloat4_b dtype_traits specializations (block-float data is unpacked to float32 before any ndarray is built).

Privatizing the two allocate_* functions is deferred (needs a public uninitialized-allocation API plus ttnn/test migration).
