+++ tt_metal/impl/tensor/host_tensor_factory.cpp
@@ -13,7 +13,8 @@
 #include <tt-metalium/mesh_coord.hpp>
 #include <tt-metalium/experimental/tensor/spec/tensor_spec.hpp>
 #include <tt-metalium/experimental/tensor/topology/tensor_topology.hpp>
-#include <tt-metalium/experimental/tensor/impl/tensor_impl.hpp>
+
+#include "tensor_impl_private.hpp"
 
 #include <tt_stl/span.hpp>
 #include <tt_stl/fmt.hpp>
